ci: pin down block number for mainnet tests for now
Some tests started failing after the price of LUSD dipped below 1 in the
Curve pool. For now, pin the tests to a block where the price is >1.

In the long run, we should fix the failing test cases to be able to work
with any starting LUSD price, at which point we should unpin the block.